PCam2d: Stop errors on null in follow_targets.
Because the phantom camera updates its position in the editor, if you
are in FollowMode.GROUP and you go to add a node to follow_targets, upon
clicking "+ Add Element" the camera will try and get the global position
of the new null element (as you have not had time to set it yet) causing
errors to be output every frame. Now null elements will simply be
skipped.