Update PhantomCamera2D to be able to use new TileMapLayer nodes as a limit target. (#378)
* Support TileMapLayer (as Node2D) for PhantomCamera2D limit.

TileMap has been deprecated in Godot 4.3 in favour of the new
TileMapLayer. The interface is mostly the same so they can share code
paths in most cases. For backwards compatibility with Godot 4.2,
TileMapLayer type is not being used explicitly.

* Check tile_map.tile_set for null in update_limit_all_sides.

There is an issue with the new TileMapLayer type that is causing it's
connect signal to be called after the TileSet has been freed when
get_tree().quit() is called. This check is to prevent a crash.
See: https://github.com/godotengine/godot/issues/96898